Update kenzato_download_template.py
Sometimes, users upload cards with no group. This causes the script to fail with an AttributeError when it tries to fetch the group tag. Added a try/except to accept AttributeError and add the file to a "failed.txt" list to inspect. This is temporary, I'd like to make it download the card regardless to put into the "other" category. 

This means I also added a "failed.txt" path in the globals.